fix(framework): adjust setDirectPrice for decimals in simulations
Problem: setDirectPrice does not directly set the returned value
of the getPrice function. Rather, it is implied that the price is
scaled by 1e18. Thus, when we call getPrice, extra zeros get
appended to the price for tokens with < 18 decimals.

Solution: Do a reverse scaling, i.e. divide by 10**(18 - decimals)