refactor: get vTokens before registering the test suite
Problem: In tests we want to iterate over the vTokens and use a
separate it() clause for each vToken. This way we can see whether
an individual vToken passes our checks or not.

Solution: Use the facto that our forking() clause registers the
tests asynchronously and, hence, allows await inside of it. Get
the vTokens first, and then register the corresponding test suite
for each vToken.